requirements.txt
torch
pandas
numpy
scikit-learn
pyyaml